@@ -15,13 +15,13 @@
 apply plugin: 'com.android.application'
 
 android {
-    compileSdkVersion 23
+    compileSdkVersion 29
 
     defaultConfig {
-        minSdkVersion 7
-        targetSdkVersion 23
-        versionCode 17
-        versionName "17"
+        minSdkVersion 14
+        targetSdkVersion 29
+        versionCode 18
+        versionName "18"
     }
     buildTypes {
         release {
@@ -29,10 +29,13 @@ android {
             shrinkResources true
             proguardFiles getDefaultProguardFile('proguard-android.txt'), 'proguard-rules.txt'
         }
+        lintOptions {
+            abortOnError false
+        }
     }
 }
 
 dependencies {
-    implementation 'com.android.support:appcompat-v7:23.1.1'
+    implementation 'androidx.appcompat:appcompat:1.0.0'
     implementation fileTree(dir: 'libs', include: ['*.jar'])
 }
